app/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
mysql-connector-python==8.0.32
databases==0.9.0
SQLAlchemy==2.0.36
python-dotenv==1.0.1
pydantic==2.10.3
pydantic_core==2.27.1
pytz==2024.2
PyJWT==2.10.1